+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 ### next
+  * tpm2_loadexternal: RSA Public and Private PEM support.
+  * tpm2_loadexternal: Object Attribute support.
+  * tpm2_loadexternal: Object authorization support.
+  * tpm2_loadexternal: Default hierarchy changes to the *null* hierarchy.
   * tpm2_verifysignature: stop outputting message hash.
   * tpm2_verifysignature: issues a warning when ticket is specified for a NULL hierarchy.
   * tpm2_verifysignature: make -t optional.